Seperate Server from App
